Split ParsingHelpers classes into their own files
#397
