+++ src/Microsoft.AspNet.Http.Extensions/Internal/HeaderSegment.cs
@@ -1,0 +1,62 @@
+﻿using System;
+
+namespace Microsoft.AspNet.Http.Internal
+{
+    internal struct HeaderSegment : IEquatable<HeaderSegment>
+    {
+        private readonly StringSegment _formatting;
+        private readonly StringSegment _data;
+
+        // <summary>
+        // Initializes a new instance of the <see cref="T:System.Object"/> class.
+        // </summary>
+        public HeaderSegment(StringSegment formatting, StringSegment data)
+        {
+            _formatting = formatting;
+            _data = data;
+        }
+
+        public StringSegment Formatting
+        {
+            get { return _formatting; }
+        }
+
+        public StringSegment Data
+        {
+            get { return _data; }
+        }
+
+        public bool Equals(HeaderSegment other)
+        {
+            return _formatting.Equals(other._formatting) && _data.Equals(other._data);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj))
+            {
+                return false;
+            }
+
+            return obj is HeaderSegment && Equals((HeaderSegment)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            unchecked
+            {
+                return (_formatting.GetHashCode() * 397) ^ _data.GetHashCode();
+            }
+        }
+
+        public static bool operator ==(HeaderSegment left, HeaderSegment right)
+        {
+            return left.Equals(right);
+        }
+
+        public static bool operator !=(HeaderSegment left, HeaderSegment right)
+        {
+            return !left.Equals(right);
+        }
+    }
+}
